net: wireguard: Use 64-bit delta in replay-window check
wg_replay_check() computed the distance between the incoming counter and
the stored replay counter into a uint32_t, even though both counters are
64-bit. A delta larger than 2^32 was therefore truncated before being
compared against the 32-bit window size:

  - On a forward jump of 2^32 + k the truncated delta (k) looked like a
    small in-window step, so the window was shifted by k and stale bits
    were kept instead of resetting the window. Those stale bits alias
    unrelated counters and can make a never-seen packet look like a
    replay.
  - On a backward gap of 2^32 + k an ancient counter looked in-window
    and could be accepted instead of rejected as too old.

Widen the delta to uint64_t so the comparison against the window size
uses the full distance. The bitmap shift amounts are only evaluated when
the delta is below the window size, so they remain in range.

Add replay-window unit tests covering a forward jump and a backward gap
larger than 2^32; both fail with the truncated delta and pass with the
64-bit one.

Tested with native_sim: tests/net/lib/wireguard.

(cherry picked from commit 3dc794a272468519057dffcf0f52ab9f021ce96b)

+++ subsys/net/lib/wireguard/wg_crypto.c
@@ -226,7 +226,7 @@ ZTESTABLE_STATIC bool wg_replay_check(uint32_t *replay_bitmap, uint64_t *replay_
 	 */
 	size_t replay_window_size = sizeof(*replay_bitmap) * CHAR_BIT; /* 32 bits */
 	bool ret = false;
-	uint32_t diff;
+	uint64_t diff;
 
 	/* WireGuard data packet counter starts from 0 but algorithm expects
 	 * packet numbers to start from 1.
